backend/requirements.txt
asgiref==3.5.2
Django==3.2.15
pytz==2022.2.1
sqlparse==0.4.2
django-colorfield==0.7.2
Pillow==9.2.0
djoser
djangorestframework-simplejwt==4.7.2
django-filter==22.1